backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openpyxl==3.1.2
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4
shapely==2.0.2
simplekml==1.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
